[T1-479] feat: Table Suspense적용
