+++ packages/wms/src/features/user/ui/index.ts
@@ -1,2 +1,3 @@
 export * from "./create-user-modal";
 export * from "./user-control-panel";
+export * from "./user-table";
