+++ packages/wms/src/features/user/ui/user-table/index.ts
@@ -1,0 +1,1 @@
+export * from "./UserTable";
